+++ db/migrate/20200618144228_add_feature_flags_to_users.rb
@@ -1,0 +1,14 @@
+class AddFeatureFlagsToUsers < ActiveRecord::Migration[6.0]
+  def change
+    within_renalware_schema do
+      add_column(
+        :users,
+        :feature_flags,
+        :integer,
+        default: 0,
+        null: false,
+        comment: "OR'ed feature flag bits to enable experimental features for certain users"
+      )
+    end
+  end
+end